Search feature implemented with Redux #GSM-7
A user is now able to get data from the source, but
there is no visual feedback.

The shape of the store is representative of the data
returned by the API.

Added prettifyrc for auto-formatting.

Refactored a lot of the code for an easier
understanding, might have to refactor later.

The data is being pulled with a massive fetch,
might have to refactor later to have smaller
fetches.

Tested the resolution time of the promises and
it is around 1/4 of a second to 1/5 of a second now.